refactor(llama): move InferenceStats
